+++ CmfSimpleCmsBundle.php
@@ -4,6 +4,7 @@ namespace Symfony\Cmf\Bundle\SimpleCmsBundle;
 
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
+use Doctrine\Bundle\PHPCRBundle\DependencyInjection\Compiler\DoctrinePhpcrMappingsPass;
 
 class CmfSimpleCmsBundle extends Bundle
 {
@@ -13,6 +14,18 @@ class CmfSimpleCmsBundle extends Bundle
 
         if ($container->hasExtension('jms_di_extra')) {
             $container->getExtension('jms_di_extra')->blackListControllerFile(__DIR__ . '/Controller/PageAdminController.php');
+        }
+
+        if (class_exists('Doctrine\Bundle\PHPCRBundle\DependencyInjection\Compiler\DoctrinePhpcrMappingsPass')) {
+            $container->addCompilerPass(
+                DoctrinePhpcrMappingsPass::createXmlMappingDriver(
+                    array(
+                        realpath(__DIR__ . '/Resources/config/doctrine-model') => 'Symfony\Cmf\Bundle\SimpleCmsBundle\Model',
+                        realpath(__DIR__ . '/Resources/config/doctrine-phpcr') => 'Symfony\Cmf\Bundle\SimpleCmsBundle\Doctrine\Phpcr',
+                    ),
+                    array('cmf_simple_cms.persistence.phpcr.manager_name')
+                )
+            );
         }
     }
 }