+++ Controller/PageAdminController.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Symfony\Cmf\Bundle\SimpleCmsBundle\Controller;
-
-use Sonata\AdminBundle\Controller\CRUDController as Controller;
-
-class PageAdminController extends Controller
-{
-}