+++ Doctrine/Phpcr/Page.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Symfony\Cmf\Bundle\SimpleCmsBundle\Doctrine\Phpcr;
+
+use Symfony\Cmf\Bundle\SimpleCmsBundle\Model\Page as ModelPage;
+
+/**
+ * {@inheritDoc}
+ */
+class Page extends ModelPage
+{
+    public $node;
+}