+++ Tests/Resources/app/config/config.php
@@ -4,3 +4,4 @@ $loader->import(CMF_TEST_CONFIG_DIR.'/default.php');
 $loader->import(CMF_TEST_CONFIG_DIR.'/phpcr_odm.php');
 $loader->import(CMF_TEST_CONFIG_DIR.'/sonata_admin.php');
 $loader->import(__DIR__.'/cmf_simple_cms.xml');
+$loader->import(__DIR__.'/cmf_core.yml');