+++ Tests/Resources/app/config/routing.php
@@ -6,8 +6,8 @@ $collection = new RouteCollection();
 $collection->addCollection(
     $loader->import(CMF_TEST_CONFIG_DIR.'/routing/sonata_routing.yml')
 );
-//$collection->addCollection(
-//    $loader->import(__DIR__.'/routing/my_test_routing.yml')
-//);
+$collection->addCollection(
+    $loader->import(__DIR__.'/routing/test.yml')
+);
 
 return $collection;